:sparkles: feat(NumberService): 수익률 출력 기능 구현
- 수익률 계산
  - 총 수익 계산
- 수익률 출력
  - 소수점 둘째 자리에서 반올림
- Rank의 prize 자료형 long으로 변환

+++ src/main/java/lotto/domain/Rank.java
@@ -11,9 +11,9 @@ public enum Rank {
     FIRST_PLACE(6, 2_000_000_000);
 
     private final int matchNumber;
-    private final int prize;
+    private final long prize;
 
-    Rank(int matchNumber, int prize) {
+    Rank(int matchNumber, long prize) {
         this.matchNumber = matchNumber;
         this.prize = prize;
     }
@@ -22,7 +22,7 @@ public enum Rank {
         return matchNumber;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return prize;
     }
